Supprime fichier forms dans la recherche
